Change seed to int

+++ src/uk/ac/cam/cl/bravo/PlanetBuilder/Vertex.java
@@ -31,8 +31,8 @@ class Vertex {
         originalY = y;
         originalZ = z;
 
-	    heightNoise = Noise.noise(originalX, originalY, originalZ, TIME, (int) WorldOptions.getInstance().getSeed() + Seeds.HeightSeed);
-	    climateNoise = Noise.noise(originalX, originalY, originalZ, TIME, (int) WorldOptions.getInstance().getSeed() + Seeds.ClimateSeed);
+	    heightNoise = Noise.noise(originalX, originalY, originalZ, TIME, WorldOptions.getInstance().getSeed() + Seeds.HeightSeed);
+	    climateNoise = Noise.noise(originalX, originalY, originalZ, TIME, WorldOptions.getInstance().getSeed() + Seeds.ClimateSeed);
     }
 
     public void Normalize(){
